+++ src/core/PostChatContext.ts
@@ -1,0 +1,5 @@
+export default interface PostChatContext {
+  formsProLocale: string;
+  participantJoined: boolean;
+  surveyInviteLink: string;
+}